% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_fasta.R
\name{abif_fasta}
\alias{abif_fasta}
\title{abif_fasta}
\usage{
abif_fasta(folder = NULL, exclude = NULL, trim = TRUE,
  trim.check = FALSE, export.check = TRUE, show.prog = TRUE,
  output = "V3-V6seq.FASTA")
}
\arguments{
\item{folder}{name of folder to where ab1 files are kept; need full path;
if path uses backslashes, must use double backslashes \code{\\} instead;
must be in quotes.}

\item{exclude}{name of files that you wish to exclude from fasta file;
must be in quotes, extention required, must be in a vector;
example: \code{c('file1.ab1','file2.ab1')}}

\item{trim}{logical, default TRUE; do you want to trim N's from beginning
and end of the sequence? trim determined by evaluating each nucleotide in
the sequence to see if it it is an N or not. It will then look for when
the number of non-N nucleotides is twice the number of Ns (meaning we
are well into our gene sequence) and the trim to be the most recent N
found. Both beginning and ends of the sequence are trimmed this way.}

\item{trim.check}{logical; default FALSE; print sequence before and after
trim to manually check the trim}

\item{export.check}{logical, default TRUE; export sequences before and after
trimming for convenience when examining trim job; exports as
\code{ouput_check.csv}}

\item{show.prog}{logical, default TRUE; Reports a message on what file is
currently being converted}

\item{ouput}{default is \code{'V3-V6seq.FASTA'}; name of output file;
fasta extension required; will be a FASTA file; will save in location
of folder provided; can set to FALSE if no output desired}
}
\value{
fasta file
}
\description{
Function to read Applied Biosystem Inc. Format (Ab1, ABIF) files and put
into one FASTA files for easy blasting on NCBI
}
