% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcabx.R
\name{calcabx}
\alias{calcabx}
\title{calcabx}
\usage{
calcabx(std_max, std_MIC_dil, fw_MIC_dil, sample_wt, solvent_vol)
}
\arguments{
\item{std_max}{numeric. highest abx concentration indilution series tested}

\item{std_MIC_dil}{numeric. number of dilutions to reach abx MIC}

\item{fw_MIC_dil}{numeric. number of dilutions to reach fecal water MIC}

\item{sample_wt}{numeric. sample weight in grams}

\item{solvent_vol}{numeric. volume of solvent used to maek fecal slurry -- mL}
}
\value{
concentration fo antibiltic in sample. in ug abx / g poop
}
\description{
calculate antibiotic concentration based on MIC of reference curve and MIC
of sample curve
}
