% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbymultiple.R
\name{genbymultiple}
\alias{genbymultiple}
\title{genbymultiple}
\usage{
genbymultiple(x, n, multiple)
}
\arguments{
\item{x}{numeric; starting value}

\item{n}{numeric; number of values to be in sequence}

\item{multiple}{numeric; factor by which sequence is generated}
}
\value{
vector of numbers increasing by a multiple
}
\description{
genterate sequence by multiples
}
