% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthCurve.R
\name{growthCurve}
\alias{growthCurve}
\title{growthCurve}
\usage{
growthCurve(pdata, well_include = NULL, x = "minute", y = "adj",
  colour = "abxlevel", group = "abxlevel", preview = TRUE)
}
\arguments{
\item{pdata}{dataframe. MIC data with metadata, adjusted absorbance,
and percent survival}

\item{well_include}{string or vector of string. Default NULL. Which wells to
include in analysis}

\item{x}{string. Default \code{'minute'}. x axis of growth curve.}

\item{y}{string. Default \code{'adj'}. y axis of growth curve.}

\item{colour}{string. Default \code{'abxlevel'}. characteristic to colour code by.}

\item{group}{string. Default \code{'abxlevel'}. characteristic to group curves by}

\item{preview}{logic. Default \code{TRUE}. Preview plot}
}
\value{
growth curve. ggplot.
}
\description{
plotting growth curve of MIC plate assay
}
