% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVictor.R
\name{importVictor}
\alias{importVictor}
\title{importVictor}
\usage{
importVictor(file)
}
\arguments{
\item{file}{path and file name}
}
\value{
list of:
    df - dataframe of wells, absorbance. is the first sheet of the victor file
    plate - 2nd sheet of victor file
    protocol - 3rd sheet of victor file
    errors - 4th sheet of victor file
    notes - 5th sheet of victor file
}
\description{
import raw data from victor
}
