% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_blast.R
\name{parse_blast}
\alias{parse_blast}
\title{parse_blast}
\usage{
parse_blast(filename = NULL, ngroups = 1, tophit = FALSE,
  output = "blast_result.csv")
}
\arguments{
\item{filename}{name of xml file containing blast search results for
V3klV6r sanger sequencing (exported from ncbi blast); full path to
file should be included}

\item{ngroups}{default 1; the number of identity-coverage groups to be
returned}

\item{tophit}{default FALSE; if TRUE, only returns the top hit for each query}

\item{output}{default 'blast_result.csv'; name of output file; will be in
csv format; will be saved in same location as input file.}
}
\value{
csv file with blast results
}
\description{
Function to parse xml output of ncbi blast of V3kl V6r regions
returns query cover, evalue, percent identity, matched organism(s),
accession number, query sequence
returns hits in percent identity-coverage groups, starting from the group
with the highest scores in both those measures, and in decreasing order
}
