% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_biplot.R
\name{pca_biplot}
\alias{pca_biplot}
\title{pca_biplot}
\usage{
pca_biplot(data, metadata, colour = NULL, label = TRUE,
  loadings = TRUE, loadings.label = TRUE,
  loadings.label.colour = "darkred")
}
\arguments{
\item{data}{dataframe. Independant variable in rows,
dependant variable in columns. Variable names in row and column names}

\item{metadata}{dataframe. metadata for the independant variables
rownames must match rownames in data}

\item{colour}{string. column name in metadata to colour code score}

\item{label}{logic. default TRUE. label score points}

\item{loadings}{logic. default TRUE. When set, show loadings on biplot}

\item{loadings.label}{logic. default TRUE. When set, label loadings}

\item{loadings.label.colour}{string. default 'darkred'}
}
\value{
list of:
        \code{p_biplot} ggplot biplot
        \code{pcx} prcomp object.
        \code{pcx_summary} dataframe. summary of pca
}
\description{
uses ggfortify and ggplot2's autoplot to make biplot.
uses prcomp to calculate pca
}
