% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_score.R
\name{pca_score}
\alias{pca_score}
\title{Functions for plotting pca results in score plot}
\usage{
pca_score(pca_data, PCs = 1:2, colour = NULL, shape = NULL,
  size = NULL, label = NULL, title = NULL, legend.ori = "vertical",
  pt.size = 3, outline = FALSE, CIgroup = NULL)
}
\arguments{
\item{pca_data}{the PCA model output of prcomp()}

\item{PCs}{numeric; vector of Principal Components to consider}

\item{colour}{string; name of column to be used as the colour aesthetic}

\item{shape}{string; name of column to be used as the shape aesthetic}

\item{size}{string; name of column to be used as the size aesthetic}

\item{label}{string; name of column to be used as the point labels}

\item{pt.size}{numeric; default 3; size of points}

\item{CIgroup'}{draws an ellipse based on covariance matrix at the
95% confidence level around the specified group
(one of the metadata columns)
specifies the groupings around which ellipses should be drawn}
}
\value{
list of: \code{p} ggplot object with generated plot
                 \code{collected_data} all the data used to plot pca data
                     using ggplot2. can use collected_data for
                     further customization using ggplot
}
\description{
Functions for plotting pca results in score plot
}
