% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Performing PCA analysis}
\usage{
perform_pca(data, sampleID = "sample_name", variable, value = "value",
  scale = "UV", center = TRUE, convert.NA = FALSE, print = TRUE)
}
\arguments{
\item{data}{dataframe; data to be used for PCA. Sample metadata is in columns
and the variables involved (i.e. compound or species)}

\item{sampleID}{string. column name for sampleID. Default \code{'sample_name'}}

\item{variable}{string; the column name in which variables are stored}

\item{value}{string; default 'value'; the column name in which data values
are stored (i.e. 'concentration' or 'abundance')}

\item{scale}{'UV' for unit variance, 'UVN' for no unit variance, 'pareto';
default UV}

\item{center}{logical; mean centering; default to TRUE}

\item{convert.NA}{logical; where value is NA, convert value to 0;
desirable when want to explicitly say missing values are true zeros
(i.e. below limit of detection) and not missing from analysis}

\item{print}{logic. Default \code{TRUE}. print preview of pca results}
}
\value{
prcomp object; PCA-processed data
}
\description{
Performing PCA analysis
}
