% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalCurve.R
\name{survivalCurve}
\alias{survivalCurve}
\title{survivalCurve}
\usage{
survivalCurve(pdata, well_include = NULL,
  timepoint = max(pdata$Repeat), x = "wellconc", y = "survival",
  preview = TRUE)
}
\arguments{
\item{pdata}{dataframe. MIC data with metadata, adjusted absorbance,
and percent survival}

\item{well_include}{string or vector of string. Default NULL. Which wells to
include in analysis}

\item{timepoint}{Default \code{99}. Which timepoint in the growth curve
should the survival curve be made from. Should correspond to the
\code{Repeat} value in the reading. If multiple timepoints provided,
produces survival curve for each timepoint.}

\item{x}{string. Default \code{'minute'}. x axis of growth curve.}

\item{y}{string. Default \code{'survival'}. y axis of growth curve.}

\item{preview}{logic. Defulat \code{TRUE}. Preview plot}
}
\value{
growth curve. ggplot.
}
\description{
plotting survival curve of MIC plate assay
}
