% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvival.R
\name{calcSurvival}
\alias{calcSurvival}
\title{calcSurvival}
\usage{
calcSurvival(full, ctrl_by = "row", ctrl_label = "growthctrl")
}
\arguments{
\item{full}{dataframe. A datafram with all your metadata, absorbance,
adjusted absorbance information}

\item{ctrl_by}{\code{c('row', 'col')} Default 'row'
\code{'row'} means growth control wells are in the same row as corresponding wells
\code{'col'} means growth control wells are in the same column as corresponding wells}

\item{ctrl_label}{string. Default \code{'growthctrl'}.
How are growth control wells identified in metadata}
}
\value{
dataframe. full, with percent survival in survival column.
}
\description{
Calculate percent survival using growth control well as 100%
}
