% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCtrl.R
\name{checkCtrl}
\alias{checkCtrl}
\title{checkCtrl}
\usage{
checkCtrl(plateDF = NULL, metadata = NULL, full = NULL,
  well_include = NULL, blank_by = "row", x = "minute", y = "abs",
  colour = "abxlevel", group = "abxlevel")
}
\arguments{
\item{plateDF}{dataframe. first sheet of victor}

\item{metadata}{dataframe. well metadata must have columns:
Well, platerow, platecol, strain, curveID, media, abx, abxlevel}

\item{well_include}{string or vector of string. Default NULL. Which wells to
include in analysis}

\item{blank_by}{\code{c('row','col')}. Default \code{'row'}.
Sets orientation of blanks.
\code{'row'} means blanks are in the same row as corresponding wells
\code{'col'} means blanks are in the same column as corresponding wells}

\item{x}{string. x-axis of plot. default \code{'minute'}}

\item{y}{string. y-axis of plot. default \code{'abs'}}

\item{colour}{string. characteristic to colour by on plot. Default 'abxlevel'}

\item{group}{string. characteristic to colour by on plot. Default 'abxlevel'}
}
\value{
ggplot. growth curve of the negatice and postive control
}
\description{
check negative and positive control of MIC assay
}
