% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCtrl.R
\name{checkCtrl}
\alias{checkCtrl}
\title{checkCtrl}
\usage{
checkCtrl(plateDF, metadata, well_include = NULL, x = "minute",
  y = "abs", colour = "wellconc", group = "wellconc")
}
\arguments{
\item{plateDF}{dataframe. first sheet of victor}

\item{metadata}{dataframe. well metadata must have columns:
Well, platerow, platecol, strain, curveID, media, abx, wellconc}

\item{well_include}{string or vector of string. Default NULL. Which wells to
include in analysis}

\item{x}{string. x-axis of plot. default \code{'minute'}}

\item{y}{string. y-axis of plot. default \code{'abs'}}

\item{colour}{string. characteristic to colour by on plot. Default 'wellconc'}

\item{group}{string. characteristic to colour by on plot. Default 'wellconc'}
}
\value{
ggplot. growth curve of the negative and postive control
}
\description{
check negative and positive control of MIC assay
}
