% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVictor.R
\name{importVictor}
\alias{importVictor}
\title{importVictor}
\usage{
importVictor(file)
}
\arguments{
\item{file}{path and file name

\code{df} - dataframe of wells, absorbance. is the first sheet of the victor file;}
}
\description{
import raw data from victor
}
